Blake2, Part 1-3/5: Extract with key support

+++ tests/curve64-ours.c
@@ -46,7 +46,7 @@ main()
     return EXIT_SUCCESS;
 
   bool ok = true;
-  for (int i = 0; i < sizeof(vectors) / sizeof(curve25519_test_vector); ++i) {
+  for (size_t i = 0; i < sizeof(vectors) / sizeof(curve25519_test_vector); ++i) {
     ok &= print_test(vectors[i].scalar, vectors[i].public, vectors[i].secret);
   }
